A new meta-model - Actress meta-model inspired by EMF
In the end we need a full meta-model. This is a lighter version of EMF meta-model. Right now there is only an immutable implementation based on case classes. This is not very good and it should be updated to support proxies. Also ActressPackage is currently created manually, but it would be much better to be able to create it using reflection at runtime so the specification, i.e., the A* classes really reflect the resulting instances.