Improvements to the model and new draft of a protocol.
There is no binding. The parent is responsible for its binding and passes an instance to its child actor do its business. Actors are deployed following containments. It should scale well as the actors do not need any state.
There is still a problem with the meta-model implementation in the fact that it is using a singleton registry (just like EMF) and that does not play well with actors. We simply need a registry that is bound to an actor implementation, perhaps it can be implemented using agents. The thing is that each instance has to have its _actor set.